Upgrade mongodb to version 5
Also migrate the dependency to a peer dependency, so the user can choose what mongodb version to use. For the sake of this project, no version specifics seem to be used. So it does not really matter what version we get, as long as it is mongodb 4 or 5.

(cherry picked from commit e4d944680db7222482ce55340eaddf15c02c234d)